requirements.txt
customtkinter==5.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0